pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lit-agent"
version = "0.1.0"
description = "Reference extraction agent for analyzing Deepsearch results"
authors = [{name = "Research Team"}]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.79.1",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers --cov=lit_agent --cov-report=term-missing"
markers = [
  "unit: fast, isolated tests",
  "integration: tests that hit real services or I/O"
]

[tool.coverage.run]
branch = true
source = ["src/lit_agent"]

[tool.coverage.report]
skip_empty = true

